Rework the rules interface for queries.
Instead of requiring the caller to construct and error-check the query
arguments up front, cache an encoding error in the Query and report it when the
query is invoked. This is mainly ergonomic, but I think it makes the API easier
to use and read.

In addition, simplify the structure of the query arguments to reduce stutter.
Instead of constructor functions, Set as an interface and make the arguments
slices that implement that interface.

Update tests.