Add a size-limiting transport wrapper.
When recording, we need to limit the size of response bodies from the server so
that streaming methods do not stall the recorder, which reads the entire
response body before returning any data to the caller.

To work around this, a limitTransport wraps the http.RoundTripper we pass to
the recorder, to artificially truncate the response body to a fixed limit.

This limit is only applied when recording, and the default maximum size was
chosen empirically: You want as small a limit as possible so as not to blow up
the test data size, but if it's too small the client will fail spuriously.

This unblocks the manual tests:

- Drop the -manual flag.
- Rename the file.
- Make the tests fail during recording without a limit set.
- Update testdata.