Add package ostatus to handle tweet creation and deletion.
Twitter API v2 does not (yet) support creating or deleting tweets.  This
package uses the v1.1 API to support those operations, but replies are
translated into the v2 message format in a rudimentary way.

Eventually I would like to have a proper "status" package, but I am saving that
name for a future where the v2 API has its own methods for this.