Remove the old generated code before generating.
If the old generated code does not compile, the generator will not be able to
build since it links the base package. However, the rest of the package is fine
without the generated code, so by unlinking it ahead of time we ensure the tool
will be able to start.