Move API-nonspecific client parts to new package jhttp.
Most of the client plumbing is not specific to the Twitter API.  There are some
compatible assumptions (e.g., comma-separated parameters), but most of it is
just a skin around net/http. Pull this out into a separate package.

The layer graph is now:

   1. http types
   2. twitter
   3. users tweets auth ostatus

Most of the auth package could be lifted further up the graph, but for now I've
left it alone.